requirements.txt
PyMySQL
mysql-replication